Задание Модуль5-задание 1
